producer-service/requirements.txt
Flask==2.3.3
kafka-python==2.0.2
mysql-connector-python==8.1.0
